agent-sdk/requirements.txt
# Claude Agent SDK
anthropic==0.40.0

# Vector search
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# Hybrid search (BM25 + Vector)
rank-bm25==0.2.2
jieba==0.42.1

# Database
sqlite-vec==0.1.6

# Web framework (for local API)
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.3

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
Pillow==10.2.0
python-pptx==1.0.2
PyMuPDF==1.24.0
matplotlib==3.9.0
chardet==5.2.0

# Computer Use
pyautogui==0.9.54
pyperclip==1.9.0
pytesseract==0.3.13

# Browser Use
playwright==1.48.0

# Others
requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

# Web Search (UAPI)
uapi-sdk-python==0.1.5
